+++ src/yukimonsai/tactical/ShipIconImpl.java
@@ -5,6 +5,9 @@ import com.fs.starfarer.api.combat.*;
 import com.fs.starfarer.api.fleet.FleetMemberAPI;
 import com.fs.starfarer.api.graphics.SpriteAPI;
 import com.fs.starfarer.api.input.InputEventAPI;
+import com.fs.starfarer.combat.CombatFleetManager;
+import com.fs.starfarer.combat.tasks.CombatTaskManager;
+import lunalib.backend.util.ReflectionUtils;
 import org.lwjgl.util.vector.Vector2f;
 import org.magiclib.util.MagicUI;
 
@@ -15,6 +18,11 @@ public class ShipIconImpl implements ShipIcon {
     DeployedFleetMemberAPI member;
     ShipAPI ship;
 
+    // updated when DataUpdate is true
+    float armor_l = 0f;
+    float armor_r = 0f;
+    float armor_u = 0f;
+    float armor_d = 0f;
 
     public static Color ESCORT_COLOR = new Color(75, 253, 63);
 
@@ -25,6 +33,8 @@ public class ShipIconImpl implements ShipIcon {
     public static Color YOU_COLOR = new Color(255, 255, 255);
     public static Color SND_COLOR = new Color(250, 39, 190);
     public static Color RETREAT_COLOR = new Color(250, 222, 39);
+    public static float ARMOR_THRESH = 50;
+    public static float ascalemult = 1.25f;
 
 
     public ShipIconImpl(DeployedFleetMemberAPI member) {
@@ -71,11 +81,12 @@ public class ShipIconImpl implements ShipIcon {
     }
 
     @Override
-    public void render(boolean flip, boolean flipv, float XX, float YY, float w, float h, HashMap<String, CombatFleetManagerAPI.AssignmentInfo> assignmentList, boolean withText, float sineAmt) {
+    public void render(boolean flip, boolean flipv, float XX, float YY, float w, float h, HashMap<String, CombatFleetManagerAPI.AssignmentInfo> assignmentList, boolean withText, float sineAmt, boolean UIUpdate, boolean DataUpdate) {
         // icon
+        ship = member.getShip();
         int side = ship.getOwner();
         SpriteAPI sprite = Global.getSettings().getSprite(member.getMember().getHullSpec().getSpriteName());
-        float hp = member.getShip().getHullLevel();
+        float hp = ship.getHullLevel();
         float colorScale = 200;
 
         Color color = NA_SettingsListener.na_combatui_colorblind ?
@@ -90,17 +101,17 @@ public class ShipIconImpl implements ShipIcon {
         //sprite.setCenter(sprite.getWidth(), sprite.getHeight());
 
 
-        boolean retreating = Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(member.getShip()) != null
-                && Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(member.getShip()).getType().equals(CombatAssignmentType.RETREAT);
+        boolean retreating = Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(ship) != null
+                && Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(ship).getType().equals(CombatAssignmentType.RETREAT);
 
-        boolean escort = !NA_SettingsListener.na_combatui_info && !retreating && Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(member.getShip()) != null
+        boolean escort = !NA_SettingsListener.na_combatui_info && !retreating && Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(ship) != null
                 && (
-                Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(member.getShip()).getType().equals(CombatAssignmentType.LIGHT_ESCORT)
-                        || Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(member.getShip()).getType().equals(CombatAssignmentType.MEDIUM_ESCORT)
-                        || Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(member.getShip()).getType().equals(CombatAssignmentType.HEAVY_ESCORT)
+                Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(ship).getType().equals(CombatAssignmentType.LIGHT_ESCORT)
+                        || Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(ship).getType().equals(CombatAssignmentType.MEDIUM_ESCORT)
+                        || Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(ship).getType().equals(CombatAssignmentType.HEAVY_ESCORT)
         );
-        boolean snd = !NA_SettingsListener.na_combatui_info && Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(member.getShip()) != null
-                && Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(member.getShip()).getType().equals(CombatAssignmentType.SEARCH_AND_DESTROY);
+        boolean snd = !NA_SettingsListener.na_combatui_info && Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(ship) != null
+                && Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(ship).getType().equals(CombatAssignmentType.SEARCH_AND_DESTROY);
 
         if (retreating) {
             sprite.setAlphaMult(0.6f + 0.39f * (float)sineAmt);
@@ -113,16 +124,16 @@ public class ShipIconImpl implements ShipIcon {
 
 
         float yyy = 0;
-        if (member.getShip() == Global.getCombatEngine().getPlayerShip()) {
-            MagicUI.addText(member.getShip(), "you", YOU_COLOR, new Vector2f(XX + 6, YY + h + yyy), false);
+        if (ship == Global.getCombatEngine().getPlayerShip()) {
+            MagicUI.addText(ship, "you", YOU_COLOR, new Vector2f(XX + 6, YY + h + yyy), false);
             yyy -= 9;
         } else if (Global.getCombatEngine().getPlayerShip() != null && Global.getCombatEngine().getPlayerShip().getShipTarget() != null
-                && Global.getCombatEngine().getPlayerShip().getShipTarget().getId().equals(member.getShip().getId())) {
+                && Global.getCombatEngine().getPlayerShip().getShipTarget().getId().equals(ship.getId())) {
 
             SpriteAPI select = Global.getSettings().getSprite("icons","na_icon_select");
             select.setSize(w, w);
-            if (side == 1) select.setColor(new Color(145, 2, 2, 255));
-            else select.setColor(new Color(50, NA_SettingsListener.na_combatui_colorblind ? 175 : 255, NA_SettingsListener.na_combatui_colorblind ? 255 : 50, 255));
+            if (side == 1) select.setColor(new Color(145, 40, 2, 255));
+            else select.setColor(new Color(170, NA_SettingsListener.na_combatui_colorblind ? 175 : 255, NA_SettingsListener.na_combatui_colorblind ? 255 : 50, 255));
             select.setAdditiveBlend();
             select.setAlphaMult(0.7f + 0.25f * sineAmt);
             select.renderAtCenter(XX + w * 0.5f, YY + h * 0.5f);
@@ -130,36 +141,36 @@ public class ShipIconImpl implements ShipIcon {
         }
         if (side == 0 || !NA_SettingsListener.na_combatui_noenemyinfo) {
             if (snd) {
-                MagicUI.addText(member.getShip(), "S&D", SND_COLOR, new Vector2f(XX + 6, YY + h + yyy), false);
+                MagicUI.addText(ship, "S&D", SND_COLOR, new Vector2f(XX + 6, YY + h + yyy), false);
                 yyy -= 9;
             }
-            if (!NA_SettingsListener.na_combatui_info && assignmentList.containsKey(member.getShip().getId())) {
-                if (assignmentList.get(member.getShip().getId()).getAssignedMembers().isEmpty()) {
-                    MagicUI.addText(member.getShip(), "-", ESCORT_COLOR, new Vector2f(XX + 6, YY + h + yyy), false);
+            if (!NA_SettingsListener.na_combatui_info && assignmentList.containsKey(ship.getId())) {
+                if (assignmentList.get(ship.getId()).getAssignedMembers().isEmpty()) {
+                    MagicUI.addText(ship, "-", ESCORT_COLOR, new Vector2f(XX + 6, YY + h + yyy), false);
                 } else
-                if (assignmentList.get(member.getShip().getId()).getType().equals(CombatAssignmentType.MEDIUM_ESCORT))
-                    MagicUI.addText(member.getShip(), "M", ESCORT_COLOR, new Vector2f(XX + 6, YY + h + yyy), false);
-                else if (assignmentList.get(member.getShip().getId()).getType().equals(CombatAssignmentType.HEAVY_ESCORT))
-                    MagicUI.addText(member.getShip(), "H", ESCORT_COLOR, new Vector2f(XX + 6, YY + h + yyy), false);
-                else if (assignmentList.get(member.getShip().getId()).getType().equals(CombatAssignmentType.LIGHT_ESCORT))
-                    MagicUI.addText(member.getShip(), "L", ESCORT_COLOR, new Vector2f(XX + 6, YY + h + yyy), false);
+                if (assignmentList.get(ship.getId()).getType().equals(CombatAssignmentType.MEDIUM_ESCORT))
+                    MagicUI.addText(ship, "M", ESCORT_COLOR, new Vector2f(XX + 6, YY + h + yyy), false);
+                else if (assignmentList.get(ship.getId()).getType().equals(CombatAssignmentType.HEAVY_ESCORT))
+                    MagicUI.addText(ship, "H", ESCORT_COLOR, new Vector2f(XX + 6, YY + h + yyy), false);
+                else if (assignmentList.get(ship.getId()).getType().equals(CombatAssignmentType.LIGHT_ESCORT))
+                    MagicUI.addText(ship, "L", ESCORT_COLOR, new Vector2f(XX + 6, YY + h + yyy), false);
                 yyy -= 9;
             } else if (!NA_SettingsListener.na_combatui_info) {
-                if (Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(member.getShip()) != null
-                        && Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(member.getShip()).getType() == CombatAssignmentType.DEFEND) {
-                    MagicUI.addText(member.getShip(), "DEF", D_COLOR, new Vector2f(XX + 6, YY + h + yyy), false);
+                if (Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(ship) != null
+                        && Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(ship).getType() == CombatAssignmentType.DEFEND) {
+                    MagicUI.addText(ship, "DEF", D_COLOR, new Vector2f(XX + 6, YY + h + yyy), false);
                     yyy -= 9;
-                } else if (Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(member.getShip()) != null
-                        && Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(member.getShip()).getType() == CombatAssignmentType.CAPTURE) {
-                    MagicUI.addText(member.getShip(), "CAP", CAP_COLOR, new Vector2f(XX + 6, YY + h + yyy), false);
+                } else if (Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(ship) != null
+                        && Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(ship).getType() == CombatAssignmentType.CAPTURE) {
+                    MagicUI.addText(ship, "CAP", CAP_COLOR, new Vector2f(XX + 6, YY + h + yyy), false);
                     yyy -= 9;
-                } else if (Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(member.getShip()) != null
-                        && Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(member.getShip()).getType() == CombatAssignmentType.CONTROL) {
-                    MagicUI.addText(member.getShip(), "CON", D_COLOR, new Vector2f(XX + 6, YY + h + yyy), false);
+                } else if (Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(ship) != null
+                        && Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(ship).getType() == CombatAssignmentType.CONTROL) {
+                    MagicUI.addText(ship, "CON", D_COLOR, new Vector2f(XX + 6, YY + h + yyy), false);
                     yyy -= 9;
-                } else if (Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(member.getShip()) != null
-                        && Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(member.getShip()).getType() == CombatAssignmentType.ENGAGE) {
-                    MagicUI.addText(member.getShip(), "eng", ENGAGE_COLOR, new Vector2f(XX + 6, YY + h + yyy), false);
+                } else if (Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(ship) != null
+                        && Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(ship).getType() == CombatAssignmentType.ENGAGE) {
+                    MagicUI.addText(ship, "eng", ENGAGE_COLOR, new Vector2f(XX + 6, YY + h + yyy), false);
                     yyy -= 9;
                 }
 
@@ -171,33 +182,33 @@ public class ShipIconImpl implements ShipIcon {
         if ((side == 0 || !NA_SettingsListener.na_combatui_noenemyinfo) && !snd && escort) {
 
 
-            if (Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(member.getShip()).getType().equals(CombatAssignmentType.LIGHT_ESCORT)
-                    || Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(member.getShip()).getType().equals(CombatAssignmentType.MEDIUM_ESCORT)
-                    || Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(member.getShip()).getType().equals(CombatAssignmentType.HEAVY_ESCORT)) {
-                MagicUI.addText(member.getShip(), "esc", ESCORT_COLOR, new Vector2f(XX + 6, YY + h + yyy), false);
+            if (Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(ship).getType().equals(CombatAssignmentType.LIGHT_ESCORT)
+                    || Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(ship).getType().equals(CombatAssignmentType.MEDIUM_ESCORT)
+                    || Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(ship).getType().equals(CombatAssignmentType.HEAVY_ESCORT)) {
+                MagicUI.addText(ship, "esc", ESCORT_COLOR, new Vector2f(XX + 6, YY + h + yyy), false);
                 yyy -= 9;
             }
         }
 
         if (side == 0 || !NA_SettingsListener.na_combatui_noenemyinfo) {
             if (!NA_SettingsListener.na_combatui_info && retreating) {
-                MagicUI.addText(member.getShip(), (NA_SettingsListener.na_combatui_hspace + NA_SettingsListener.na_combatui_size < 55) ? ((NA_SettingsListener.na_combatui_hspace + NA_SettingsListener.na_combatui_size <= 46) ? "ret." : "retr.") : "retreat", RETREAT_COLOR, new Vector2f(XX + 6, YY + h + yyy), false);
+                MagicUI.addText(ship, (NA_SettingsListener.na_combatui_hspace + NA_SettingsListener.na_combatui_size < 55) ? ((NA_SettingsListener.na_combatui_hspace + NA_SettingsListener.na_combatui_size <= 46) ? "ret." : "retr.") : "retreat", RETREAT_COLOR, new Vector2f(XX + 6, YY + h + yyy), false);
                 yyy -= 9;
             } else if (!NA_SettingsListener.na_combatui_info) {
-                if (sineAmt > 0 && member.getShip() != Global.getCombatEngine().getPlayerShip()
-                        && ((member.getShip().getShipAI() != null && member.getShip().getShipAI().getAIFlags() != null && (
-                        member.getShip().getShipAI().getAIFlags().hasFlag(ShipwideAIFlags.AIFlags.IN_CRITICAL_DPS_DANGER)
-                                || member.getShip().getShipAI().getAIFlags().hasFlag(ShipwideAIFlags.AIFlags.HAS_INCOMING_DAMAGE)
-                                || member.getShip().getShipAI().getAIFlags().hasFlag(ShipwideAIFlags.AIFlags.RUN_QUICKLY)
-                                || member.getShip().getShipAI().getAIFlags().hasFlag(ShipwideAIFlags.AIFlags.TURN_QUICKLY)
-                                || member.getShip().getShipAI().getAIFlags().hasFlag(ShipwideAIFlags.AIFlags.BACKING_OFF)
+                if (sineAmt > 0 && ship != Global.getCombatEngine().getPlayerShip()
+                        && ((ship.getShipAI() != null && ship.getShipAI().getAIFlags() != null && (
+                        ship.getShipAI().getAIFlags().hasFlag(ShipwideAIFlags.AIFlags.IN_CRITICAL_DPS_DANGER)
+                                || ship.getShipAI().getAIFlags().hasFlag(ShipwideAIFlags.AIFlags.HAS_INCOMING_DAMAGE)
+                                || ship.getShipAI().getAIFlags().hasFlag(ShipwideAIFlags.AIFlags.RUN_QUICKLY)
+                                || ship.getShipAI().getAIFlags().hasFlag(ShipwideAIFlags.AIFlags.TURN_QUICKLY)
+                                || ship.getShipAI().getAIFlags().hasFlag(ShipwideAIFlags.AIFlags.BACKING_OFF)
                 )))) {
-                    if (member.getShip().getShipAI().getAIFlags().hasFlag(ShipwideAIFlags.AIFlags.IN_CRITICAL_DPS_DANGER)
-                            || member.getShip().getShipAI().getAIFlags().hasFlag(ShipwideAIFlags.AIFlags.HAS_INCOMING_DAMAGE)
-                            || member.getShip().getShipAI().getAIFlags().hasFlag(ShipwideAIFlags.AIFlags.RUN_QUICKLY)
-                            || member.getShip().getShipAI().getAIFlags().hasFlag(ShipwideAIFlags.AIFlags.TURN_QUICKLY))
-                        MagicUI.addText(member.getShip(), "!!!", ALERT_COLOR, new Vector2f(XX + 6, YY + h + yyy), false);
-                    else MagicUI.addText(member.getShip(), "!", ENGAGE_COLOR, new Vector2f(XX + 6, YY + h + yyy), false);
+                    if (ship.getShipAI().getAIFlags().hasFlag(ShipwideAIFlags.AIFlags.IN_CRITICAL_DPS_DANGER)
+                            || ship.getShipAI().getAIFlags().hasFlag(ShipwideAIFlags.AIFlags.HAS_INCOMING_DAMAGE)
+                            || ship.getShipAI().getAIFlags().hasFlag(ShipwideAIFlags.AIFlags.RUN_QUICKLY)
+                            || ship.getShipAI().getAIFlags().hasFlag(ShipwideAIFlags.AIFlags.TURN_QUICKLY))
+                        MagicUI.addText(ship, "!!!", ALERT_COLOR, new Vector2f(XX + 6, YY + h + yyy), false);
+                    else MagicUI.addText(ship, "!", ENGAGE_COLOR, new Vector2f(XX + 6, YY + h + yyy), false);
                     yyy -= 9;
                 }
             }
@@ -205,23 +216,136 @@ public class ShipIconImpl implements ShipIcon {
 
         // bars
         float off = 0;
+        float SPACE = 9;
 
-        if (((side == 0 && !NA_SettingsListener.na_combatui_ppt) || (side == 1 && !NA_SettingsListener.na_combatui_noenemyppt)) && !retreating) {
-            Color fill2 = new Color(202, 197, 197, member.getShip().getPeakTimeRemaining() < 1f ? (int) (200 + 50 * sineAmt) : 255);
-            Color border2 = new Color(169, 232, 8, 180);
-            float crTimeFrac = member.getShip().getPeakTimeRemaining()/
-                    (1f + member.getMember().getStats().getPeakCRDuration().computeEffective(member.getShip().getHullSpec().getNoCRLossTime()));
+        if (((side == 0 && !NA_SettingsListener.na_combatui_armor) || (side == 1 && !NA_SettingsListener.na_combatui_noenemyarmor))) {
+            if (ship.getArmorGrid() != null && ship.getArmorGrid().getArmorRating() > ARMOR_THRESH) {
+                // armor
 
-            MagicUI.addBar(member.getShip(), member.getShip().getCurrentCR(), fill2, border2, crTimeFrac * member.getShip().getCurrentCR(), new Vector2f(XX + w * 0.125f, YY - 12), 6, w*0.75f, true);
-        } else off = -8;
+
+                float armor = ship.getArmorGrid().getArmorRating();
+                float armorThresh = 1f + Math.max(ARMOR_THRESH,  armor*.35f);
+
+                if (DataUpdate) {
+                    updateArmor();
+                }
+
+
+
+                if (armor_l > 0) {
+                    float ar = Math.max(0, Math.min(1f, (armor_l-armorThresh)/(armor - armorThresh)));
+                    Color armorColor = NA_SettingsListener.na_combatui_colorblind ?
+                            new Color(250 - (int)(colorScale * 0.8 * ar), 50 + (int)(0.7f * colorScale * ar), 50 + (int)(colorScale * ar), 220)
+                            : new Color(250 - (int)(colorScale * 0.9 * ar), 50 + (int)(colorScale * ar), 50, 220);
+
+                    sprite = Global.getSettings().getSprite("icons", "ytd_armor_l");
+                    sprite.setSize(w, h);
+                    sprite.setColor(armorColor);
+                    if (armor_l < armorThresh)
+                        sprite.setAlphaMult(0.59f - 0.4f * sineAmt);
+                    else sprite.setAlphaMult(NA_SettingsListener.na_combatui_armorAlpha);
+                    sprite.setNormalBlend();
+                    sprite.renderAtCenter(XX + w * 0.5f, YY + h * 0.5f);
+                    sprite.setAdditiveBlend();
+                    sprite.renderAtCenter(XX + w * 0.5f, YY + h * 0.5f);
+                }
+                if (armor_r > 0) {
+                    float ar = Math.max(0, Math.min(1f, (armor_r-armorThresh)/(armor - armorThresh)));
+                    Color armorColor = NA_SettingsListener.na_combatui_colorblind ?
+                            new Color(250 - (int)(colorScale * 0.8 * ar), 50 + (int)(0.7f * colorScale * ar), 50 + (int)(colorScale * ar), 220)
+                            : new Color(250 - (int)(colorScale * 0.9 * ar), 50 + (int)(colorScale * ar), 50, 220);
+
+                    sprite = Global.getSettings().getSprite("icons", "ytd_armor_r");
+                    sprite.setSize(w, h);
+                    sprite.setColor(armorColor);
+                    if (armor_r < armorThresh)
+                        sprite.setAlphaMult(0.59f - 0.4f * sineAmt);
+                    else sprite.setAlphaMult(NA_SettingsListener.na_combatui_armorAlpha);
+                    sprite.setNormalBlend();
+                    sprite.renderAtCenter(XX + w * 0.5f, YY + h * 0.5f);
+                    sprite.setAdditiveBlend();
+                    sprite.renderAtCenter(XX + w * 0.5f, YY + h * 0.5f);
+                }
+                if (armor_u > 0) {
+                    float ar = Math.max(0, Math.min(1f, (armor_u-armorThresh)/(armor - armorThresh)));
+                    Color armorColor = NA_SettingsListener.na_combatui_colorblind ?
+                            new Color(250 - (int)(colorScale * 1.1 * ar), 25 + (int)(0.8f * colorScale * ar), 25 + (int)(1.1*colorScale * ar), 240)
+                            : new Color(250 - (int)(colorScale * 1.1 * ar), 25 + (int)(1.1*colorScale * ar), 25, 240);
+
+                    sprite = Global.getSettings().getSprite("icons", "ytd_armor_u");
+                    sprite.setSize(w, h);
+                    sprite.setColor(armorColor);
+                    if (armor_u < armorThresh)
+                        sprite.setAlphaMult(0.59f - 0.4f * sineAmt);
+                    else sprite.setAlphaMult(NA_SettingsListener.na_combatui_armorAlpha);
+                    sprite.setNormalBlend();
+                    sprite.renderAtCenter(XX + w * 0.5f, YY + h * 0.5f);
+                    sprite.setAdditiveBlend();
+                    sprite.renderAtCenter(XX + w * 0.5f, YY + h * 0.5f);
+                }
+                if (armor_d > 0) {
+                    float ar = Math.max(0, Math.min(1f, (armor_d-armorThresh)/(armor - armorThresh)));
+                    Color armorColor = NA_SettingsListener.na_combatui_colorblind ?
+                            new Color(250 - (int)(colorScale * 0.8 * ar), 50 + (int)(0.7f * colorScale * ar), 50 + (int)(colorScale * ar), 220)
+                            : new Color(250 - (int)(colorScale * 0.9 * ar), 50 + (int)(colorScale * ar), 50, 220);
+
+                    sprite = Global.getSettings().getSprite("icons", "ytd_armor_d");
+                    sprite.setSize(w, h);
+                    sprite.setColor(armorColor);
+                    if (armor_d < armorThresh)
+                        sprite.setAlphaMult(0.59f - 0.4f * sineAmt);
+                    else sprite.setAlphaMult(NA_SettingsListener.na_combatui_armorAlpha);
+                    sprite.setNormalBlend();
+                    sprite.renderAtCenter(XX + w * 0.5f, YY + h * 0.5f);
+                    sprite.setAdditiveBlend();
+                    sprite.renderAtCenter(XX + w * 0.5f, YY + h * 0.5f);
+                }
+
+            }
+        }
+
+        if (((side == 0 && !NA_SettingsListener.na_combatui_ppt) || (side == 1 && !NA_SettingsListener.na_combatui_noenemyppt))) {
+
+            if (!retreating) {
+                // dont show bars, do show armor though
+                float cr = ship.getCurrentCR();
+                float crMinorMal = Global.getSettings().getCRPlugin().getMalfunctionThreshold(ship.getMutableStats());
+                float crMinorMaj = Global.getSettings().getCRPlugin().getCriticalMalfunctionThreshold(ship.getMutableStats());
+
+                Color fill2 = new Color(
+                        202,
+                        ship.getPeakTimeRemaining() < 1 ?
+                                (
+                                        cr > crMinorMal ? 200 :
+                                                (cr > crMinorMaj ? 150 : 50)
+                                ) : 197,
+                        ship.getPeakTimeRemaining() < 1 ? (
+                                cr > crMinorMaj ? 100 : 50
+                        )
+                                : 197, ship.getPeakTimeRemaining() < 1f ? (int) (200 + 50 * sineAmt) : 255);
+                Color border2 = new Color(169, 232, 8, 0); // was 180
+                float crTimeFrac = ship.getPeakTimeRemaining()/
+                        (1f + member.getMember().getStats().getPeakCRDuration().computeEffective(ship.getHullSpec().getNoCRLossTime()));
+
+                MagicUI.addBar(ship, crTimeFrac > 0 ? ship.getCurrentCR() : cr, fill2, border2, 0, new Vector2f(XX + w * 0.125f, YY - 3 - SPACE), 4, w*0.75f, false);
+            }
+
+        } else off = -SPACE;
 
         if ((side == 0 && !NA_SettingsListener.na_combatui_flux) || (side == 1 && !NA_SettingsListener.na_combatui_noenemyflux)) {
-            Color fill = new Color(210, member.getShip().getFluxTracker().isOverloaded() ? 150 : 82, 237, member.getShip().getFluxTracker().isOverloadedOrVenting() ? (int) (200 + 50 * sineAmt) : 255);
+            Color fill = new Color(210, ship.getFluxTracker().isOverloaded() ? 150 : 82, 237, ship.getFluxTracker().isOverloadedOrVenting() ? (int) (200 + 50 * sineAmt) : 255);
             Color border = new Color(175, 134, 227, 180);
 
-            MagicUI.addBar(member.getShip(), member.getShip().getFluxLevel(), fill, border, member.getShip().getHardFluxLevel(), new Vector2f(XX + w * 0.125f, YY - 4 + off), 6, w * 0.75f, true);
+            MagicUI.addBar(ship, ship.getFluxLevel(), fill, border, ship.getHardFluxLevel(), new Vector2f(XX + w * 0.125f, YY - 4 + off), 6, w * 0.75f, true);
 
         }
+    }
+
+    public void updateArmor() {
+        armor_l = ship.getAverageArmorInSlice(ship.getFacing() + 90f, 60);
+        armor_r = ship.getAverageArmorInSlice(ship.getFacing() - 90f, 60);
+        armor_u = ship.getAverageArmorInSlice(ship.getFacing(), 60);
+        armor_d = ship.getAverageArmorInSlice(ship.getFacing() + 180f, 60);
     }
 
     @Override
@@ -236,56 +360,67 @@ public class ShipIconImpl implements ShipIcon {
             if (side == 0) {
                 if (commandMode == NA_CombatPlugin.CommandMode.ESCORT_COMMAND) {
                     CombatAssignmentType escortType = CombatAssignmentType.LIGHT_ESCORT;
-                    if (member.getShip().getHullSize() == ShipAPI.HullSize.CAPITAL_SHIP) escortType = CombatAssignmentType.HEAVY_ESCORT;
-                    else if (member.getShip().getHullSize() == ShipAPI.HullSize.CRUISER) escortType = CombatAssignmentType.MEDIUM_ESCORT;
+                    if (ship.getHullSize() == ShipAPI.HullSize.CAPITAL_SHIP) escortType = CombatAssignmentType.HEAVY_ESCORT;
+                    else if (ship.getHullSize() == ShipAPI.HullSize.CRUISER) escortType = CombatAssignmentType.MEDIUM_ESCORT;
 
-                    if (!assignmentList.containsKey(member.getShip().getId())) {
+                    if (!assignmentList.containsKey(ship.getId())) {
 
-                        if (Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(member.getShip()) != null) {
+                        if (Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(ship) != null) {
                             Global.getCombatEngine().getFleetManager(side).getTaskManager(false).removeAssignment(
-                                    Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(member.getShip()));
+                                    Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(ship));
                         }
 
 
                         Global.getCombatEngine().getFleetManager(side).getTaskManager(false).createAssignment(escortType, member, true);
                         Global.getSoundPlayer().playUISound("ui_button_full_retreat", 1f, 1f);
-                    } else if (assignmentList.containsKey(member.getShip().getId())) {
+                    } else if (assignmentList.containsKey(ship.getId())) {
 
                         Global.getCombatEngine().getFleetManager(side).getTaskManager(false).removeAssignment(
-                                assignmentList.get(member.getShip().getId()));
+                                assignmentList.get(ship.getId()));
                         Global.getSoundPlayer().playUISound("ui_button_full_retreat", 1f, 1f);
                     }
                 } else if (commandMode == NA_CombatPlugin.CommandMode.SEARCHANDDESTROY_COMMAND) {
                     //List<CombatFleetManagerAPI.AssignmentInfo> assignments = Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAllAssignments();
-                    if (Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(member.getShip()) == null
-                            || !Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(member.getShip()).getType().equals(CombatAssignmentType.SEARCH_AND_DESTROY)) {
+                    if (Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(ship) == null
+                            || !Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(ship).getType().equals(CombatAssignmentType.SEARCH_AND_DESTROY)) {
 
                         Global.getCombatEngine().getFleetManager(side).getTaskManager(false).orderSearchAndDestroy(member, true);
                         Global.getSoundPlayer().playUISound("ui_button_full_retreat", 1f, 1f);
+                    } else if (Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(ship) != null
+                            && Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(ship).getType().equals(CombatAssignmentType.SEARCH_AND_DESTROY)) {
+
+
+                        Object[] arr = new Object[1]; arr[0] = member;
+                        ReflectionUtils.invoke("cancelDirectOrdersForMember", Global.getCombatEngine().getFleetManager(side).getTaskManager(false), arr, null, 1);
+
+
+                        Global.getSoundPlayer().playUISound("ui_button_full_retreat", 1f, 1f);
                     }
                 } else if (commandMode == NA_CombatPlugin.CommandMode.RETREAT_COMMAND) {
-                    if (Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(member.getShip()) == null
-                            || !Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(member.getShip()).getType().equals(CombatAssignmentType.RETREAT)) {
+                    if (Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(ship) == null
+                            || !Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(ship).getType().equals(CombatAssignmentType.RETREAT)) {
 
-                        if (Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(member.getShip()) != null) {
+                        if (Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(ship) != null) {
                             Global.getCombatEngine().getFleetManager(side).getTaskManager(false).removeAssignment(
-                                    Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(member.getShip()));
+                                    Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(ship));
                         }
 
 
-                        Global.getCombatEngine().getFleetManager(side).getTaskManager(false).orderRetreat(member, true, true);
+                        Global.getCombatEngine().getFleetManager(side).getTaskManager(false).orderRetreat(member, true, false);
                         Global.getSoundPlayer().playUISound("ui_button_full_retreat", 1f, 1f);
-                    } /*else if (Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(member.getShip()) != null
-                            && Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(member.getShip()).getType().equals(CombatAssignmentType.RETREAT)) {
+                    } else if (Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(ship) != null
+                            && Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(ship).getType().equals(CombatAssignmentType.RETREAT)) {
 
-                        Global.getCombatEngine().getFleetManager(side).getTaskManager(false).removeAssignment(
-                                Global.getCombatEngine().getFleetManager(side).getTaskManager(false).getAssignmentFor(member.getShip()));
+
+                        Object[] arr = new Object[1]; arr[0] = member;
+                        ReflectionUtils.invoke("cancelDirectOrdersForMember", Global.getCombatEngine().getFleetManager(side).getTaskManager(false), arr, null, 1);
+
 
                         Global.getSoundPlayer().playUISound("ui_button_full_retreat", 1f, 1f);
-                    }*/
-                } else if (Global.getCombatEngine().getPlayerShip() != null && member.getShip() != Global.getCombatEngine().getPlayerShip()) {
-                    if (member.getShip() != Global.getCombatEngine().getPlayerShip().getShipTarget()) {
-                        Global.getCombatEngine().getPlayerShip().setShipTarget(member.getShip());
+                    }
+                } else if (Global.getCombatEngine().getPlayerShip() != null && ship != Global.getCombatEngine().getPlayerShip()) {
+                    if (ship != Global.getCombatEngine().getPlayerShip().getShipTarget()) {
+                        Global.getCombatEngine().getPlayerShip().setShipTarget(ship);
                         Global.getSoundPlayer().playUISound("ui_button_full_retreat", 1f, 1f);
                     } else {
                         Global.getCombatEngine().getPlayerShip().setShipTarget(null);
@@ -294,9 +429,9 @@ public class ShipIconImpl implements ShipIcon {
 
                 }
 
-            } else if (side == 1 && Global.getCombatEngine().getPlayerShip() != null && member.getShip() != Global.getCombatEngine().getPlayerShip()) {
-                if (member.getShip() != Global.getCombatEngine().getPlayerShip().getShipTarget()) {
-                    Global.getCombatEngine().getPlayerShip().setShipTarget(member.getShip());
+            } else if (side == 1 && Global.getCombatEngine().getPlayerShip() != null && ship != Global.getCombatEngine().getPlayerShip()) {
+                if (ship != Global.getCombatEngine().getPlayerShip().getShipTarget()) {
+                    Global.getCombatEngine().getPlayerShip().setShipTarget(ship);
                     Global.getSoundPlayer().playUISound("ui_button_full_retreat", 1f, 1f);
                 } else {
                     Global.getCombatEngine().getPlayerShip().setShipTarget(null);